Split computation and crypto

+++ contracts/ink/computation/lib.rs
@@ -14,13 +14,13 @@ pub mod computation {
 
         #[ink(message)]
         pub fn triangle_number(&self, n: i64) {
-            let _res: i64 = (1..=n as i64).fold(0, |sum, x| sum.wrapping_add(x));
+            let _res: i64 = (1..=n).fold(0, |sum, x| sum.wrapping_add(x));
         }
 
         #[ink(message)]
         #[allow(clippy::arithmetic_side_effects)]
         pub fn odd_product(&self, n: i32) -> i64 {
-            (1..=n as i64).fold(1, |prod, x| prod.wrapping_mul(2 * x - 1))
+            (1..=n).fold(1, |prod, x| prod.wrapping_mul(2 * x as i64 - 1))
         }
     }
 }
